Mock authentication and set environment
