+++ aspnetcore/test/integration-tests/samples/2.x/IntegrationTestsSample/tests/RazorPagesProject.Tests/IntegrationTests/BasicTests.cs
@@ -1,13 +1,6 @@
-﻿using System.Collections.Generic;
-using System.Net;
-using System.Threading.Tasks;
-using AngleSharp.Dom.Html;
+﻿using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc.Testing;
-using Microsoft.AspNetCore.TestHost;
-using Microsoft.Extensions.DependencyInjection;
 using Xunit;
-using RazorPagesProject.Services;
-using RazorPagesProject.Tests.Helpers;
 
 namespace RazorPagesProject.Tests.IntegrationTests
 {
@@ -41,78 +34,6 @@ namespace RazorPagesProject.Tests.IntegrationTests
             Assert.Equal("text/html; charset=utf-8", 
                 response.Content.Headers.ContentType.ToString());
         }
-    #endregion
-
-        #region snippet2
-        [Fact]
-        public async Task Get_SecurePageRequiresAnAuthenticatedUser()
-        {
-            // Arrange
-            var client = _factory.CreateClient(
-                new WebApplicationFactoryClientOptions
-                {
-                    AllowAutoRedirect = false
-                });
-
-            // Act
-            var response = await client.GetAsync("/SecurePage");
-
-            // Assert
-            Assert.Equal(HttpStatusCode.Redirect, response.StatusCode);
-            Assert.StartsWith("http://localhost/Identity/Account/Login", 
-                response.Headers.Location.OriginalString);
-        }
-        #endregion
-
-        #region snippet3
-        [Fact]
-        public async Task CanGetAGithubUser()
-        {
-            // Arrange
-            void ConfigureTestServices(IServiceCollection services) =>
-                services.AddSingleton<IGithubClient>(new TestGithubClient());
-            var client = _factory
-                .WithWebHostBuilder(builder => 
-                    builder.ConfigureTestServices(ConfigureTestServices))
-                .CreateClient();
-
-            // Act
-            var profile = await client.GetAsync("/GithubProfile");
-            Assert.Equal(HttpStatusCode.OK, profile.StatusCode);
-            var profileHtml = await HtmlHelpers.GetDocumentAsync(profile);
-
-            var profileWithUserName = await client.SendAsync(
-                (IHtmlFormElement)profileHtml.QuerySelector("#user-profile"),
-                new Dictionary<string, string> { ["Input_UserName"] = "user" });
-
-            // Assert
-            Assert.Equal(HttpStatusCode.OK, profileWithUserName.StatusCode);
-            var profileWithUserHtml = 
-                await HtmlHelpers.GetDocumentAsync(profileWithUserName);
-            var userLogin = profileWithUserHtml.QuerySelector("#user-login");
-            Assert.Equal("user", userLogin.TextContent);
-        }
-
-        public class TestGithubClient : IGithubClient
-        {
-            public Task<GithubUser> GetUserAsync(string userName)
-            {
-                if (userName == "user")
-                {
-                    return Task.FromResult(
-                        new GithubUser
-                        {
-                            Login = "user",
-                            Company = "Contoso Blockchain",
-                            Name = "John Doe"
-                        });
-                }
-                else
-                {
-                    return Task.FromResult<GithubUser>(null);
-                }
-            }
-        }
-        #endregion
     }
+    #endregion
 }
